refactor: organização das rotas em uma pasta services e correção do tipo date nos models
