+++ src/main.py
@@ -1,33 +1,15 @@
-import datetime as dt
 from models import *
 from utils.generate_tables import generate_tables
-from utils.generate_id import generate_id
-import csv
+from services.configs import headers, storage_directory
+from services.users import router as users_router
+from services.messages import router as messages_router
+from services.products import router as products_router
+from services.favorites import router as favorites_router
+from services.sales import router as sales_router
 
 from fastapi import FastAPI
 
 app = FastAPI()
-
-# Define os cabeçalhos para cada tabela
-headers = {
-    "users": ["id", "name", "year", "cpf", "gender", "phone_number", "address", "email", "password"],
-    "messages": ["id", "user_sent_id", "user_received_id", "title", "description", "created_at"],
-    "products": ["id", "seller_id", "title", "description", "price", "quantity", "image_url"],
-    "favorites": ["id", "user_id", "product_id"],
-    "sales": ["id", "seller_id", "buyer_id", "product_id", "created_at"]
-};
-
-# Nomeando diretório de armazenamento dos dados
-storage_directory = "data";
-
-# Nomeando caminho para os arquivos CSV
-path_directories = {
-    "users": f"{storage_directory}/users.csv",
-    "messages": f"{storage_directory}/messages.csv",
-    "products": f"{storage_directory}/products.csv",
-    "favorites": f"{storage_directory}/favorites.csv",
-    "sales": f"{storage_directory}/sales.csv"
-}
 
 # Inicializando arquivos CSV
 generate_tables(storage_directory, headers);
@@ -37,450 +19,17 @@ generate_tables(storage_directory, headers);
 def read_root():
     return {"message": "Seja bem vindo a Lojita"}
 
-# Rota para criar um novo usuário
-@app.post("/users")
-async def create_user(user: User):
-    try:
-        # Escrever os dados no arquivo CSV
-        with open(path_directories["users"], mode="a", newline="", encoding="utf-8") as file:
-            writer = csv.DictWriter(file, fieldnames=headers["users"]);
-            
-            # Gerando id aleatório
-            user.id = generate_id(16);
-            
-            # Adicionar a linha com os dados do usuário
-            writer.writerow(dict(user));
-        return {"message": "User created successfully", "data": user};
-    except Exception as e:
-        return {"error": str(e)}
-    
-@app.get("/users/{id}")
-async def get_user(id: str):
-    try:
-        with open(path_directories["users"], mode="r", newline="", encoding="utf-8") as file:
-            reader = csv.DictReader(file);
-            for row in reader:
-                if row["id"] == id:
-                    return {"user": row};
-            return {"message": "User not found"};
-    except Exception as e:
-        return {"error": str(e)}
+# Incluindo rotas de usuários
+app.include_router(users_router);
 
-@app.put("/users/{id}")
-async def update_user(id: str, user: User):
-    try:
-        with open(path_directories["users"], mode="r", newline="", encoding="utf-8") as file:
-            reader = csv.DictReader(file);
-            rows = list(reader);
-            updated = False;
-            
-            for row in rows:
-                if row["id"] == id:
-                    row["name"] = user.name;
-                    row["year"] = user.year;
-                    row["cpf"] = user.cpf;
-                    row["gender"] = user.gender;
-                    row["phone_number"] = user.phone_number;
-                    row["address"] = user.address;
-                    row["email"] = user.email;
-                    row["password"] = user.password;
-                    
-                    updated = True;
-                    break;
-            
-            if updated:
-                with open(path_directories["users"], mode="w", newline="", encoding="utf-8") as file:
-                    writer = csv.DictWriter(file, fieldnames=headers["users"]);
-                    writer.writeheader();
-                    writer.writerows(rows);
-                    
-                return {"message": "User updated successfully"};
-            else:
-                return {"message": "User not found"};
-    except Exception as e:
-        return {"error": str(e)}
-    
-@app.delete("/users/{id}")
-async def delete_user(id: str):
-    try:
-        with open(path_directories["users"], mode="r", newline="", encoding="utf-8") as file:
-            reader = csv.DictReader(file);
-            rows = list(reader);
-            deleted = False;
-            
-            for row in rows:
-                if row["id"] == id:
-                    rows.remove(row);
-                    deleted = True;
-                    break;
-            
-            if deleted:
-                with open(path_directories["users"], mode="w", newline="", encoding="utf-8") as file:
-                    writer = csv.DictWriter(file, fieldnames=headers["users"]);
-                    writer.writeheader();
-                    writer.writerows(rows);
-                    
-                return {"message": "User deleted successfully"};
-            else:
-                return {"message": "User not found"};
-    except Exception as e:
-        return {"error": str(e)};
-    
-@app.post("/messages")
-async def create_message(message: Message):
-    try:
-        sent_exist = False;
-        receiver_exist = False;
-        
-        with open(path_directories["users"], mode="r", newline="", encoding="utf-8") as file:
-            reader = csv.DictReader(file);
-            for row in reader:
-                if row["id"] == message.user_sent_id:
-                    sent_exist = True;
-                if row["id"] == message.user_received_id:
-                    receiver_exist = True;
-                if sent_exist and receiver_exist:
-                    break;
-        
-        if sent_exist and receiver_exist:
-            with open(path_directories["messages"], mode="a", newline="", encoding="utf-8") as file:
-                writer = csv.DictWriter(file, fieldnames=headers["messages"]);
-                
-                # Gerando id aleatório
-                message.id = generate_id(16);
-                # Adicionando data e hora do envio da mensagem
-                message.created_at = dt.datetime.now();
-                
-                # Adicionar a linha com os dados da requisição
-                writer.writerow(dict(message));
-                
-                return {"message": "Message created successfully", "data": message};
-        else:
-            return {"message": "User not found"};
-    except Exception as e:
-        return {"error": str(e)}
+# Incluindo rotas de mensagens
+app.include_router(messages_router);
 
-@app.get("/messages/{id}")
-async def get_message(id: str):
-    try:
-        with open(path_directories["messages"], mode="r", newline="", encoding="utf-8") as file:
-            reader = csv.DictReader(file);
-            for row in reader:
-                if row["id"] == id:
-                    return {"message": row};
-            return {"message": "Message not found"};
-    except Exception as e:
-        return {"error": str(e)}
-    
-@app.put("/messages/{id}")
-async def update_message(id: str, message: Message):
-    try:
-        with open(path_directories["messages"], mode="r", newline="", encoding="utf-8") as file:
-            reader = csv.DictReader(file);
-            rows = list(reader);
-            updated = False;
-            
-            for row in rows:
-                if row["id"] == id:
-                    row["title"] = message.title;
-                    row["description"] = message.description;
-                    
-                    updated = True;
-                    break;
-            
-            if updated:
-                with open(path_directories["messages"], mode="w", newline="", encoding="utf-8") as file:
-                    writer = csv.DictWriter(file, fieldnames=headers["messages"]);
-                    writer.writeheader();
-                    writer.writerows(rows);
-                    
-                return {"message": "Message updated successfully"};
-            else:
-                return {"message": "Message not found"};
-    except Exception as e:
-        return {"error": str(e)}
-    
-@app.delete("/messages/{id}")
-async def delete_message(id: str):
-    try:
-        with open(path_directories["messages"], mode="r", newline="", encoding="utf-8") as file:
-            reader = csv.DictReader(file);
-            rows = list(reader);
-            deleted = False;
-            
-            for row in rows:
-                if row["id"] == id:
-                    rows.remove(row);
-                    deleted = True;
-                    break;
-            
-            if deleted:
-                with open(path_directories["messages"], mode="w", newline="", encoding="utf-8") as file:
-                    writer = csv.DictWriter(file, fieldnames=headers["messages"]);
-                    writer.writeheader();
-                    writer.writerows(rows);
-                    
-                return {"message": "Message deleted successfully"};
-            else:
-                return {"message": "Message not found"};
-    except Exception as e:
-        return {"error": str(e)};
-    
-@app.post("/products")
-async def create_product(product: Product):
-    try:
-        seller_exist = False;
-        
-        with open(path_directories["users"], mode="r", newline="", encoding="utf-8") as file:
-            reader = csv.DictReader(file);
-            for row in reader:
-                if row["id"] == product.seller_id:
-                    seller_exist = True;
-                    break;
-        
-        if seller_exist:
-            # Escrever os dados no arquivo CSV
-            with open(path_directories["products"], mode="a", newline="", encoding="utf-8") as file:
-                writer = csv.DictWriter(file, fieldnames=headers["products"]);
-                
-                # Gerando id aleatório
-                product.id = generate_id(16);
-                
-                # Adicionar a linha com os dados do usuário
-                writer.writerow(dict(product));
-            return {"message": "Product created successfully", "data": product};
-        else:
-            return {"message": "User not found"};
-    except Exception as e:
-        return {"error": str(e)}
-    
-@app.get("/products/{id}")
-async def get_product(id: str):
-    try:
-        with open(path_directories["products"], mode="r", newline="", encoding="utf-8") as file:
-            reader = csv.DictReader(file);
-            for row in reader:
-                if row["id"] == id:
-                    return {"product": row};
-            return {"message": "Product not found"};
-    except Exception as e:
-        return {"error": str(e)};
-    
-@app.put("/products/{id}")
-async def update_product(id: str, product: Product):
-    try:
-        with open(path_directories["products"], mode="r", newline="", encoding="utf-8") as file:
-            reader = csv.DictReader(file);
-            rows = list(reader);
-            updated = False;
-            
-            for row in rows:
-                if row["id"] == id:
-                    row["title"] = product.title;
-                    row["description"] = product.description;
-                    row["price"] = product.price;
-                    row["quantity"] = product.quantity;
-                    row["image_url"] = product.image_url;
-                    
-                    updated = True;
-                    break;
-            
-            if updated:
-                with open(path_directories["products"], mode="w", newline="", encoding="utf-8") as file:
-                    writer = csv.DictWriter(file, fieldnames=headers["products"]);
-                    writer.writeheader();
-                    writer.writerows(rows);
-                    
-                return {"message": "Product updated successfully"};
-            else:
-                return {"message": "Product not found"};
-    except Exception as e:
-        return {"error": str(e)};
-    
-@app.delete("/products/{id}")
-async def delete_product(id: str):
-    try:
-        with open(path_directories["products"], mode="r", newline="", encoding="utf-8") as file:
-            reader = csv.DictReader(file);
-            rows = list(reader);
-            deleted = False;
-            
-            for row in rows:
-                if row["id"] == id:
-                    rows.remove(row);
-                    deleted = True;
-                    break;
-            
-            if deleted:
-                with open(path_directories["products"], mode="w", newline="", encoding="utf-8") as file:
-                    writer = csv.DictWriter(file, fieldnames=headers["products"]);
-                    writer.writeheader();
-                    writer.writerows(rows);
-                    
-                return {"message": "Product deleted successfully"};
-            else:
-                return {"message": "Product not found"};
-    except Exception as e:
-        return {"error": str(e)};
-    
-@app.post("/favorites")
-async def create_favorite(favorite: Favorite):
-    try:
-        user_exist = False;
-        
-        with open(path_directories["users"], mode="r", newline="", encoding="utf-8") as file:
-            reader = csv.DictReader(file);
-            for row in reader:
-                if row["id"] == favorite.user_id:
-                    user_exist = True;
-                    break;
-                
-        product_exist = False;
-        
-        with open(path_directories["products"], mode="r", newline="", encoding="utf-8") as file:
-            reader = csv.DictReader(file);
-            for row in reader:
-                if row["id"] == favorite.product_id:
-                    product_exist = True;
-                    break;
-        
-        if user_exist and product_exist:
-            # Escrever os dados no arquivo CSV
-            with open(path_directories["favorites"], mode="a", newline="", encoding="utf-8") as file:
-                writer = csv.DictWriter(file, fieldnames=headers["favorites"]);
-                
-                # Gerando id aleatório
-                favorite.id = generate_id(16);
-                
-                # Adicionar a linha com os dados do usuário
-                writer.writerow(dict(favorite));
-            return {"message": "Favorite created successfully", "data": favorite};
-        else:
-            return {"message": "User or product not found"};
-    except Exception as e:
-        return {"error": str(e)}
-    
-@app.get("/favorites/{id}")
-async def get_favorite(id: str):
-    try:
-        with open(path_directories["favorites"], mode="r", newline="", encoding="utf-8") as file:
-            reader = csv.DictReader(file);
-            for row in reader:
-                if row["id"] == id:
-                    return {"favorite": row};
-            return {"message": "Favorite not found"};
-    except Exception as e:
-        return {"error": str(e)};
-    
-@app.delete("/favorites/{id}")
-async def delete_favorite(id: str):
-    try:
-        with open(path_directories["favorites"], mode="r", newline="", encoding="utf-8") as file:
-            reader = csv.DictReader(file);
-            rows = list(reader);
-            deleted = False;
-            
-            for row in rows:
-                if row["id"] == id:
-                    rows.remove(row);
-                    deleted = True;
-                    break;
-            
-            if deleted:
-                with open(path_directories["favorites"], mode="w", newline="", encoding="utf-8") as file:
-                    writer = csv.DictWriter(file, fieldnames=headers["favorites"]);
-                    writer.writeheader();
-                    writer.writerows(rows);
-                    
-                return {"message": "Favorite deleted successfully"};
-            else:
-                return {"message": "Favorite not found"};
-    except Exception as e:
-        return {"error": str(e)};
-    
-@app.post("/sales")
-async def create_sale(sale: Sale):
-    try:
-        seller_exist = False;
-        
-        with open(path_directories["users"], mode="r", newline="", encoding="utf-8") as file:
-            reader = csv.DictReader(file);
-            for row in reader:
-                if row["id"] == sale.seller_id:
-                    seller_exist = True;
-                    break;
-                
-        buyer_exist = False;
-        
-        with open(path_directories["users"], mode="r", newline="", encoding="utf-8") as file:
-            reader = csv.DictReader(file);
-            for row in reader:
-                if row["id"] == sale.buyer_id:
-                    buyer_exist = True;
-                    break;
-                
-        product_exist = False;
-        
-        with open(path_directories["products"], mode="r", newline="", encoding="utf-8") as file:
-            reader = csv.DictReader(file);
-            for row in reader:
-                if row["id"] == sale.product_id:
-                    product_exist = True;
-                    break;
-        
-        if seller_exist and buyer_exist and product_exist:
-            # Escrever os dados no arquivo CSV
-            with open(path_directories["sales"], mode="a", newline="", encoding="utf-8") as file:
-                writer = csv.DictWriter(file, fieldnames=headers["sales"]);
-                
-                # Gerando id aleatório
-                sale.id = generate_id(16);
-                
-                # Adicionando data e hora da compra
-                sale.created_at = dt.datetime.now();
-                
-                # Adicionar a linha com os dados do usuário
-                writer.writerow(dict(sale));
-            return {"message": "Sale created successfully", "data": sale};
-        else:
-            return {"message": "Seller, buyer or product not found"};
-    except Exception as e:
-        return {"error": str(e)}
-    
-@app.get("/sales/{id}")
-async def get_sale(id: str):
-    try:
-        with open(path_directories["sales"], mode="r", newline="", encoding="utf-8") as file:
-            reader = csv.DictReader(file);
-            for row in reader:
-                if row["id"] == id:
-                    return {"sale": row};
-            return {"message": "Sale not found"};
-    except Exception as e:
-        return {"error": str(e)};
-    
-@app.delete("/sales/{id}")
-async def delete_sale(id: str):
-    try:
-        with open(path_directories["sales"], mode="r", newline="", encoding="utf-8") as file:
-            reader = csv.DictReader(file);
-            rows = list(reader);
-            deleted = False;
-            
-            for row in rows:
-                if row["id"] == id:
-                    rows.remove(row);
-                    deleted = True;
-                    break;
-            
-            if deleted:
-                with open(path_directories["sales"], mode="w", newline="", encoding="utf-8") as file:
-                    writer = csv.DictWriter(file, fieldnames=headers["sales"]);
-                    writer.writeheader();
-                    writer.writerows(rows);
-                    
-                return {"message": "Sale deleted successfully"};
-            else:
-                return {"message": "Sale not found"};
-    except Exception as e:
-        return {"error": str(e)};
+# Incluindo rotas de produtos
+app.include_router(products_router);
+
+# Incluindo rotas de favoritos
+app.include_router(favorites_router);
+
+# Incluindo rotas de vendas
+app.include_router(sales_router);
