+++ src/services/configs.py
@@ -1,0 +1,20 @@
+# Define os cabeçalhos para cada tabela
+headers = {
+    "users": ["id", "name", "year", "cpf", "gender", "phone_number", "address", "email", "password"],
+    "messages": ["id", "user_sent_id", "user_received_id", "title", "description", "created_at"],
+    "products": ["id", "seller_id", "title", "description", "price", "quantity", "image_url"],
+    "favorites": ["id", "user_id", "product_id"],
+    "sales": ["id", "seller_id", "buyer_id", "product_id", "created_at"]
+};
+
+# Nomeando diretório de armazenamento dos dados
+storage_directory = "data";
+
+# Nomeando caminho para os arquivos CSV
+path_directories = {
+    "users": f"{storage_directory}/users.csv",
+    "messages": f"{storage_directory}/messages.csv",
+    "products": f"{storage_directory}/products.csv",
+    "favorites": f"{storage_directory}/favorites.csv",
+    "sales": f"{storage_directory}/sales.csv"
+}
